[Refactor] 컴포넌트화
