+++ my-diary/pages/write.tsx
@@ -1,101 +1,5 @@
-import { useState } from "react";
-import { useRouter } from "next/router";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { Textarea } from "@/components/ui/textarea";
-import { ArrowLeft } from "lucide-react";
+import DiaryForm from "@/components/DiaryForm";
 
 export default function WritePage() {
-  const router = useRouter();
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [error, setError] = useState("");
-
-  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    setIsSubmitting(true);
-    setError("");
-
-    const formData = new FormData(e.currentTarget);
-    const title = formData.get("title") as string;
-    const content = formData.get("content") as string;
-
-    if (!title.trim() || !content.trim()) {
-      setError("제목과 내용을 모두 입력해주세요.");
-      setIsSubmitting(false);
-      return;
-    }
-
-    try {
-      const response = await fetch("/api/diary", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ title, content }),
-      });
-
-      if (!response.ok) {
-        throw new Error("handleSubmit : 일기 저장 에러");
-      }
-
-      router.push("/");
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "handleSubmit : error");
-      setIsSubmitting(false);
-    }
-  };
-
-  return (
-    <div className="container mx-auto px-4 py-8">
-      <div className="max-w-2xl mx-auto">
-        <Button variant="ghost" className="mb-4" onClick={() => router.back()}>
-          <ArrowLeft className="mr-2 h-4 w-4" />
-          돌아가기
-        </Button>
-
-        <h1 className="text-3xl font-bold mb-8">일기 쓰기</h1>
-
-        <form onSubmit={handleSubmit} className="space-y-6">
-          <div className="space-y-2">
-            <Label htmlFor="title">제목</Label>
-            <Input
-              id="title"
-              name="title"
-              placeholder="일기 제목을 입력하세요"
-              required
-              maxLength={100}
-            />
-          </div>
-
-          <div className="space-y-2">
-            <Label htmlFor="content">내용</Label>
-            <Textarea
-              id="content"
-              name="content"
-              placeholder="하루를 기록하세요!"
-              required
-              className="min-h-[300px]"
-            />
-          </div>
-
-          {error && <p className="text-red-500 text-sm">{error}</p>}
-
-          <div className="flex justify-end gap-4">
-            <Button
-              type="button"
-              variant="outline"
-              onClick={() => router.back()}
-              disabled={isSubmitting}
-            >
-              취소
-            </Button>
-            <Button type="submit" disabled={isSubmitting}>
-              {isSubmitting ? "저장 중..." : "저장하기"}
-            </Button>
-          </div>
-        </form>
-      </div>
-    </div>
-  );
+  return <DiaryForm />;
 }
